refactor(card): remove type="button"
This removes the <Card type="button" /> variant from our <Card /> component. It's a change agreed
with design as it renders a variation of our card that looks like a button and can be conceived as
an UX anti-pattern.

BREAKING CHANGE: You won't longer be able to use the "button" variant in the "type" prop of `<Card
/>`.

21