refactor(icons): rename icons
Rename all our icons so they have the word `Icon` on their name. In this way, when importing one of
our "icon" components it's very clear that the component responsibility it to render an icon as
opposed to some other thing.

BREAKING CHANGE: Now all icon components have the word "Icon" on their name.

20