feat: add banner and logoContainer to NavbarTheme
This commit introduces two new properties to the NavbarTheme interface:
banner and logoContainer. The banner property defines the width and
height of the banner, while the logoContainer property provides
additional styling options for the logo container, including width,
height, and paddingLeft. This change allows for more flexible and
detailed theming of the navbar component.